refactor: Mode method related functions
from wakepy.core.mode to wakepy.core.method
