+++ tests/unit/test_core/test_method.py
@@ -6,8 +6,14 @@ import sys
 import pytest
 
 from wakepy.core import DBusMethodCall
-from wakepy.core.method import Method, MethodOutcome, MethodOutcomeValue
-from wakepy.core.mode import has_enter, has_exit, has_heartbeat
+from wakepy.core.method import (
+    Method,
+    MethodOutcome,
+    MethodOutcomeValue,
+    has_enter,
+    has_exit,
+    has_heartbeat,
+)
 from wakepy.core.registry import MethodRegistryError, get_method
 
 if sys.version_info < (3, 8):  # pragma: no-cover-if-py-gte-38
